Добавлен столбец средний рейтинг в таблицу товары!

+++ Advisor.Database/Tables/Products.sql
@@ -8,7 +8,7 @@
     [DateOfCreate] DATETIME2 NOT NULL, 
     [Name] NVARCHAR(100) NOT NULL, 
     [CategoryId] INT NOT NULL,
-	[Rating] INT NOT NULL, 
+	[Rating] NUMERIC(5, 5) NOT NULL, 
     constraint fk_categories_products foreign key (CategoryId) 
 	references Categories(Id) on delete cascade on update cascade,
 	constraint fk_users_products foreign key (UserId)
